+++ hutool-http/src/test/java/cn/hutool/http/test/HttpRequestTest.java
@@ -41,15 +41,11 @@ public class HttpRequestTest {
 
 	@Test
 	@Ignore
-	public void getWithParamsTest() {
+	public void toStringTest() {
 		String url = "http://gc.ditu.aliyun.com/geocoding?ccc=你好";
 		
-		HttpRequest request = HttpRequest.get(url).setEncodeUrlParams(true).body("a=乌海");
-		String body = request.execute().body();
-		Console.log(body);
-		
-//		String body2 = HttpUtil.get(url);
-//		Console.log(body2);
+		HttpRequest request = HttpRequest.get(url).body("a=乌海");
+		Console.log(request.toString());
 	}
 
 	@Test